Move preloader to a separate project
